+++ front/pages/home/contact.tsx
@@ -1,9 +1,8 @@
-import type { GetServerSideProps } from "next";
+import type { GetStaticProps } from "next";
 import { useRouter } from "next/router";
 import type { ReactElement } from "react";
 
 import { ContactForm } from "@app/components/home/ContactForm";
-import { ContactFormThankYou } from "@app/components/home/ContactFormThankYou";
 import { HeaderContentBlock } from "@app/components/home/ContentBlocks";
 import { Grid } from "@app/components/home/ContentComponents";
 import type { LandingLayoutProps } from "@app/components/home/LandingLayout";
@@ -11,17 +10,11 @@ import LandingLayout from "@app/components/home/LandingLayout";
 import { PageMetadata } from "@app/components/home/PageMetadata";
 import TrustedBy from "@app/components/home/TrustedBy";
 import UTMPageWrapper from "@app/components/UTMPageWrapper";
-import { getSession } from "@app/lib/auth";
-import { getUserFromSession } from "@app/lib/iam/session";
 
-export const getServerSideProps: GetServerSideProps = async (context) => {
-  const session = await getSession(context.req, context.res);
-  const user = await getUserFromSession(session);
-
+export const getStaticProps: GetStaticProps = async () => {
   return {
     props: {
       gtmTrackingId: process.env.NEXT_PUBLIC_GTM_TRACKING_ID ?? null,
-      isAuthenticated: !!user,
       shape: 0,
     },
   };
@@ -41,9 +34,6 @@ export default function Contact() {
     typeof router.query.headquarters_region === "string"
       ? router.query.headquarters_region
       : undefined;
-
-  // Test mode: show thank you page directly with ?testThankYou=true
-  const testThankYou = router.query.testThankYou === "true";
 
   const subtitle = companyName ? (
     <>
@@ -73,25 +63,11 @@ export default function Contact() {
         />
         <Grid>
           <div className="col-span-12 sm:col-span-12 md:col-span-12 lg:col-span-8 lg:col-start-2 xl:col-span-8 xl:col-start-2 2xl:col-start-3">
-            {testThankYou ? (
-              <ContactFormThankYou
-                firstName="Alban"
-                lastName="Music"
-                email="alban@dust.tt"
-                phone="+33612345678"
-                language="I would like my meeting to be in English 🇬🇧🇺🇸"
-                headquartersRegion="Europe"
-                companyHeadcount="101-500"
-                howToUseDust="Testing the Default.com integration"
-                isQualified={true}
-              />
-            ) : (
-              <ContactForm
-                prefillEmail={prefillEmail}
-                prefillHeadcount={prefillHeadcount}
-                prefillRegion={prefillRegion}
-              />
-            )}
+            <ContactForm
+              prefillEmail={prefillEmail}
+              prefillHeadcount={prefillHeadcount}
+              prefillRegion={prefillRegion}
+            />
           </div>
         </Grid>
         <TrustedBy />
